requirements.txt
requests
pandas
numpy
matplotlib
folium
plotly
windrose
meteostat
pyaqsapi
python-dotenv
flask
